pyproject.toml
[tool.poetry]
name = "financial-tracker"
version = "0.1.0"
description = ""
authors = ["Osepyan Alexander <a.e.osepyan@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
dash = "^2.18.1"
pandas = "^2.2.3"
plotly = "^5.24.1"
google-api-python-client = "^2.149.0"
google-auth-httplib2 = "^0.2.0"
google-auth-oauthlib = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
